tambah alert ketika stok minus saat hapus barang masuk

+++ hapus_barang_masuk.php
@@ -12,15 +12,24 @@ $data_stok_sekarang = mysqli_fetch_array($result_data_stok);
 $stok_sekarang = $data_stok_sekarang['stock'];
 
 $perubahan = $stok_sekarang - $qty;
-$sql_perubahan_stok = "UPDATE stock SET stock='$perubahan' WHERE idbarang='$idbarang'";
-$perubahan_stok = mysqli_query($conn, $sql_perubahan_stok);
+if ($perubahan >= 0) {
+    $sql_perubahan_stok = "UPDATE stock SET stock='$perubahan' WHERE idbarang='$idbarang'";
+    $perubahan_stok = mysqli_query($conn, $sql_perubahan_stok);
 
-$sql_hapus = "DELETE FROM masuk WHERE idmasuk='$idmasuk'";
-$hapus_barang = mysqli_query($conn, $sql_hapus);
+    $sql_hapus = "DELETE FROM masuk WHERE idmasuk='$idmasuk'";
+    $hapus_barang = mysqli_query($conn, $sql_hapus);
 
-mysqli_free_result($result_data_stok);
-mysqli_close($conn);
+    mysqli_free_result($result_data_stok);
+    mysqli_close($conn);
 
-header('Location: barang_masuk.php');
+    header('Location: barang_masuk.php');
+} else {
+    echo "
+        <script>
+            alert('Maaf data tidak bisa dihapus karena akan menyebabkan stok menjadi minus!');
+            window.location.href='barang_masuk.php';
+        </script>
+    ";
+}
 
 ?>
